More generic propertyWrapper
