+++ Sources/Stevenson/JiraService.swift
@@ -201,7 +201,7 @@ extension JiraService {
         let description: String
         let name: String
         let released: Bool
-        @CodableDate<YMDDateFormat>
+        @CustomCodable<YMDDate>
         var startDate: Date
 
         public init(projectId: Int, description: String, name: String, released: Bool = false, startDate: Date) {
